+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Changelog
 * The exception's `__notes__` field will now be sent as metadata if it exists
   [#340](https://github.com/bugsnag/bugsnag-python/pull/340)
 
+### Bug fixes
+* Fix the ignore class list not accounting for nested classes
+  [#342](https://github.com/bugsnag/bugsnag-python/pull/342)
+
+
 ## v4.4.0 (2023-02-21)
 
 ### Enhancements